requirements.txt
discord.py
python-dotenv
requests
pypdf
python-docx
openpyxl
python-pptx
pytz
